requirements.txt
# Python dependencies for the project

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
bcrypt==4.0.1
requests==2.31.0
scikit-learn==1.3.2
python-dotenv==1.0.0